tests: use minicore for tests/assembly
