+++ tests/assembly/simd-intrinsic-select.rs
@@ -11,7 +11,7 @@
 //@ assembly-output: emit-asm
 //@ compile-flags: --crate-type=lib -Copt-level=3 -C panic=abort
 
-#![feature(no_core, lang_items, repr_simd, intrinsics)]
+#![feature(no_core, lang_items, repr_simd)]
 #![no_core]
 #![allow(non_camel_case_types)]
 
@@ -47,9 +47,6 @@ pub struct f64x8([f64; 8]);
 
 #[repr(simd)]
 pub struct m64x8([i64; 8]);
-
-#[rustc_intrinsic]
-unsafe fn simd_select<M, V>(mask: M, a: V, b: V) -> V;
 
 // CHECK-LABEL: select_i8x16
 #[no_mangle]
